+++ CHANGELOG.md
@@ -8,6 +8,13 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- Added `azure_dns_zone` entities
+- Added `azure_resource_group|has|azure_dns_zone` relationships
+- Added `azure_dns_record_set` entities
+- Added `azure_dns_zone|has|azure_dns_record_set` relationships
+
 ## 5.1.0 - 2020-09-09
 
 ### Added